+++ Assets/Scripts/Parameters/Unit/HealthSetting.cs
@@ -5,8 +5,8 @@ namespace ActionSample.Parameters.Unit
     [CreateAssetMenu(menuName = "ActionSample/Health Setting")]
     public class HealthSetting : ScriptableObject
     {
-        public double hp;
-        public double power;
+        public float hp;
+        public float power;
     }
 
 }
